requirements.txt
flask
pymongo
requests
termcolor
python-dotenv
